# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Encryption & Decryption with AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type EncryptRequestAlgorithm =
  (typeof EncryptRequestAlgorithm)[keyof typeof EncryptRequestAlgorithm];

export const EncryptRequestAlgorithm = {
  AES: "AES",
  Caesar: "Caesar",
  Base64: "Base64",
  ROT13: "ROT13",
  Vigenere: "Vigenere",
  XOR: "XOR",
} as const;

export interface EncryptRequest {
  text: string;
  algorithm: EncryptRequestAlgorithm;
  key?: string;
  /** @nullable */
  shift?: number | null;
}

export interface EncryptResponse {
  encrypted: string;
  algorithm: string;
  /** @nullable */
  key?: string | null;
  processingTime: number;
}

export type DecryptRequestAlgorithm =
  (typeof DecryptRequestAlgorithm)[keyof typeof DecryptRequestAlgorithm];

export const DecryptRequestAlgorithm = {
  AES: "AES",
  Caesar: "Caesar",
  Base64: "Base64",
  ROT13: "ROT13",
  Vigenere: "Vigenere",
  XOR: "XOR",
} as const;

export interface DecryptRequest {
  text: string;
  algorithm: DecryptRequestAlgorithm;
  key?: string;
  /** @nullable */
  shift?: number | null;
}

export interface DecryptResponse {
  decrypted: string;
  algorithm: string;
  processingTime: number;
}

export type GenerateKeyRequestAlgorithm =
  (typeof GenerateKeyRequestAlgorithm)[keyof typeof GenerateKeyRequestAlgorithm];

export const GenerateKeyRequestAlgorithm = {
  AES: "AES",
  Caesar: "Caesar",
  Base64: "Base64",
  ROT13: "ROT13",
  Vigenere: "Vigenere",
  XOR: "XOR",
} as const;

export interface GenerateKeyRequest {
  algorithm: GenerateKeyRequestAlgorithm;
}

export interface GenerateKeyResponse {
  key: string;
  algorithm: string;
  description: string;
}

export interface EncryptionHistoryEntry {
  id: number;
  operation: string;
  algorithm: string;
  inputLength: number;
  outputLength: number;
  processingTime: number;
  createdAt: string;
}

export interface EncryptionStats {
  totalOperations: number;
  encryptCount: number;
  decryptCount: number;
  mostUsedAlgorithm: string;
  averageProcessingTime: number;
}