# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Encryption & Decryption with AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Encrypt text with specified algorithm
 */
export const EncryptTextBody = zod.object({
  text: zod.string(),
  algorithm: zod.enum(["AES", "Caesar", "Base64", "ROT13", "Vigenere", "XOR"]),
  key: zod.string().optional(),
  shift: zod.number().nullish(),
});

export const EncryptTextResponse = zod.object({
  encrypted: zod.string(),
  algorithm: zod.string(),
  key: zod.string().nullish(),
  processingTime: zod.number(),
});

/**
 * @summary Decrypt text with specified algorithm
 */
export const DecryptTextBody = zod.object({
  text: zod.string(),
  algorithm: zod.enum(["AES", "Caesar", "Base64", "ROT13", "Vigenere", "XOR"]),
  key: zod.string().optional(),
  shift: zod.number().nullish(),
});

export const DecryptTextResponse = zod.object({
  decrypted: zod.string(),
  algorithm: zod.string(),
  processingTime: zod.number(),
});

/**
 * @summary Generate encryption key for algorithm
 */
export const GenerateKeyBody = zod.object({
  algorithm: zod.enum(["AES", "Caesar", "Base64", "ROT13", "Vigenere", "XOR"]),
});

export const GenerateKeyResponse = zod.object({
  key: zod.string(),
  algorithm: zod.string(),
  description: zod.string(),
});

/**
 * @summary Get encryption/decryption history
 */
export const GetEncryptionHistoryResponseItem = zod.object({
  id: zod.number(),
  operation: zod.string(),
  algorithm: zod.string(),
  inputLength: zod.number(),
  outputLength: zod.number(),
  processingTime: zod.number(),
  createdAt: zod.coerce.date(),
});
export const GetEncryptionHistoryResponse = zod.array(
  GetEncryptionHistoryResponseItem,
);

/**
 * @summary Get encryption usage stats
 */
export const GetEncryptionStatsResponse = zod.object({
  totalOperations: zod.number(),
  encryptCount: zod.number(),
  decryptCount: zod.number(),
  mostUsedAlgorithm: zod.string(),
  averageProcessingTime: zod.number(),
});