# lib/api-zod/src/generated/types/decryptRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Encryption & Decryption with AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DecryptRequestAlgorithm } from "./decryptRequestAlgorithm";

export interface DecryptRequest {
  text: string;
  algorithm: DecryptRequestAlgorithm;
  key?: string;
  /** @nullable */
  shift?: number | null;
}