# lib/api-zod/src/generated/types/encryptRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Encryption & Decryption with AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { EncryptRequestAlgorithm } from "./encryptRequestAlgorithm";

export interface EncryptRequest {
  text: string;
  algorithm: EncryptRequestAlgorithm;
  key?: string;
  /** @nullable */
  shift?: number | null;
}