# lib/api-zod/src/generated/types/encryptionHistoryEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Encryption & Decryption with AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface EncryptionHistoryEntry {
  id: number;
  operation: string;
  algorithm: string;
  inputLength: number;
  outputLength: number;
  processingTime: number;
  createdAt: Date;
}