# lib/api-zod/src/generated/types/generateKeyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Encryption & Decryption with AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateKeyRequestAlgorithm } from "./generateKeyRequestAlgorithm";

export interface GenerateKeyRequest {
  algorithm: GenerateKeyRequestAlgorithm;
}